Update to react-intl example with app.js
