+++ src/lib/pageWithIntl.js
@@ -1,53 +1,11 @@
-/* eslint-disable no-underscore-dangle, react/prop-types */
+import hoistNonReactStatics from 'hoist-non-react-statics'
+import { injectIntl } from 'react-intl'
 
-import React from 'react'
-import { IntlProvider, addLocaleData, injectIntl } from 'react-intl'
+export const hoistStatics = higherOrderComponent => (BaseComponent) => {
+  const NewComponent = higherOrderComponent(BaseComponent)
+  hoistNonReactStatics(NewComponent, BaseComponent)
 
-// Register React Intl's locale data for the user's locale in the browser. This
-// locale data was added to the page by `pages/_document.js`. This only happens
-// once, on initial page load in the browser.
-if (typeof window !== 'undefined' && window.ReactIntlLocaleData) {
-  Object.keys(window.ReactIntlLocaleData).forEach((lang) => {
-    addLocaleData(window.ReactIntlLocaleData[lang])
-  })
+  return NewComponent
 }
 
-export default (Page) => {
-  const IntlPage = injectIntl(Page)
-
-  return class PageWithIntl extends React.Component {
-    static async getInitialProps(context) {
-      let props
-      if (typeof Page.getInitialProps === 'function') {
-        props = await Page.getInitialProps(context)
-      }
-
-      // Get the `locale` and `messages` from the request object on the server.
-      // In the browser, use the same values that the server serialized.
-      const { req } = context
-      const { locale, messages } = req || window.__NEXT_DATA__.props
-
-      // Always update the current time on page load/transition because the
-      // <IntlProvider> will be a new instance even with pushState routing.
-      const now = Date.now()
-
-      return {
-        ...props,
-        locale,
-        messages,
-        now,
-      }
-    }
-
-    render() {
-      const {
-        locale, messages, now, ...props
-      } = this.props // eslint-disable-line react/prop-types
-      return (
-        <IntlProvider initialNow={now} locale={locale} messages={messages}>
-          <IntlPage {...props} />
-        </IntlProvider>
-      )
-    }
-  }
-}
+export default hoistStatics(injectIntl)
